+++ src/types.ts
@@ -10,7 +10,7 @@ import NumericFilter, { NumericFilterArgs } from './filters/numeric_filter';
 import Scorer from './scorer/scorer';
 import TagList from './api/tag_list';
 import Toot, { SerializableToot } from './api/objects/toot';
-import { CacheKey, NonScoreWeightName, ScoreName, TagTootsCacheKey } from './enums';
+import { CacheKey, NonScoreWeightName, ScoreName, TagTootsCacheKey, DATA_SOURCES_FOR_FILTER_OPTIONS } from './enums';
 
 // Records
 export type AccountNames = Record<mastodon.v1.Account["acct"], Account>;
@@ -38,17 +38,12 @@ export type TootNumberProp = KeysOfValueType<Toot, number>;
 
 
 // Filters
+export type FilterOptionDataSource = typeof DATA_SOURCES_FOR_FILTER_OPTIONS[number];
+export type FilterOptionUserData = {[key in FilterOptionDataSource]?: number};
+export interface BooleanFilterOption extends FilterOptionUserData, ObjWithTootCount {};
+
 export type BooleanFilters = Record<BooleanFilterName, BooleanFilter>;
 export type NumericFilters = Record<TootNumberProp, NumericFilter>;
-
-export interface BooleanFilterOption extends ObjWithTootCount {
-    // TODO: these are really the wrong cache keys for the use case but it's consistent w/demo app for now
-    [TagTootsCacheKey.FAVOURITED_TAG_TOOTS]?: number;
-    [TagTootsCacheKey.PARTICIPATED_TAG_TOOTS]?: number;
-    [TagTootsCacheKey.TRENDING_TAG_TOOTS]?: number;
-    [ScoreName.FAVOURITED_ACCOUNTS]?: number,
-    // numToots: number;  // TODO: requiring this to exist is a typing error
-};
 
 export type FeedFilterSettingsSerialized = {
     booleanFilterArgs: BooleanFilterArgs[];